requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyserial==3.5
python-multipart==0.0.6
python-dotenv==1.0.0
